requirements.txt
# Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# AI/LLM Integration  
openai>=1.52.0

# Notion API
notion-client>=2.2.1

# Environment & Config
python-dotenv>=1.0.0
pydantic>=2.9.0
pydantic-settings>=2.6.0

# HTTP & Async
httpx>=0.27.0
aiofiles>=24.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-mock>=3.14.0

# Development
black>=24.0.0
flake8>=7.0.0
isort>=5.13.0

# Logging
loguru>=0.7.2